Promote applications to a config section
This commit promotes the definition of the applications dictionary to
it's own config section. This allows experiments to be defined at any
scope Ramble's config module allows.
