+++ lib/ramble/ramble/appkit.py
@@ -24,4 +24,4 @@ from ramble.spec import Spec
 import ramble.language.application_language
 from ramble.language.application_language import *
 
-from ramble.schema.applications import OUTPUT
+from ramble.schema.types import OUTPUT
